Переносит часть функций из utils.js в point.js
