+++ src/view/sorting-view.js
@@ -31,30 +31,6 @@ function createSortingTemplate() {
   );
 }
 
-/*export default class SortingView extends AbstractView {
-  #handleSortTypeChange = null;
-
-  constructor({onSortTypeChange}) {
-    super();
-    this.#handleSortTypeChange = onSortTypeChange;
-
-    this.element.addEventListener('click', this.#sortTypeChangeHandler);
-  }
-
-  get template() {
-    return createSortingTemplate();
-  }
-
-  #sortTypeChangeHandler = (evt) => {
-    if (evt.target.tagName !== 'INPUT') {
-      return;
-    }
-
-    evt.preventDefault();
-    this.#handleSortTypeChange(evt.target.value);
-  };
-}*/
-
 export default class SortingView extends AbstractView {
 
   #handleSortTypeChange = null;
